feat: Add skeleton to ride/id route for fast load
